refactor: Move Raptor Worker and Heuristics classes to internal api
