+++ src/main/java/org/opentripplanner/transit/raptor/RaptorService.java
@@ -5,8 +5,8 @@ import org.opentripplanner.transit.raptor.api.request.RaptorRequest;
 import org.opentripplanner.transit.raptor.api.response.RaptorResponse;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTransitDataProvider;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTripSchedule;
-import org.opentripplanner.transit.raptor.api.view.Worker;
 import org.opentripplanner.transit.raptor.configure.RaptorConfig;
+import org.opentripplanner.transit.raptor.rangeraptor.internalapi.Worker;
 import org.opentripplanner.transit.raptor.service.HeuristicSearchTask;
 import org.opentripplanner.transit.raptor.service.RangeRaptorDynamicSearch;
 import org.slf4j.Logger;
