+++ src/main/java/org/opentripplanner/transit/raptor/rangeraptor/RangeRaptorWorker.java
@@ -13,10 +13,10 @@ import org.opentripplanner.transit.raptor.api.transit.RaptorTransitDataProvider;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTripSchedule;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTripScheduleSearch;
 import org.opentripplanner.transit.raptor.api.transit.TransitArrival;
-import org.opentripplanner.transit.raptor.api.view.Worker;
 import org.opentripplanner.transit.raptor.rangeraptor.internalapi.RoundProvider;
 import org.opentripplanner.transit.raptor.rangeraptor.internalapi.RoutingStrategy;
 import org.opentripplanner.transit.raptor.rangeraptor.internalapi.SlackProvider;
+import org.opentripplanner.transit.raptor.rangeraptor.internalapi.Worker;
 import org.opentripplanner.transit.raptor.rangeraptor.internalapi.WorkerState;
 import org.opentripplanner.transit.raptor.rangeraptor.lifecycle.LifeCycleEventPublisher;
 import org.opentripplanner.transit.raptor.rangeraptor.transit.AccessPaths;
