+++ src/main/java/org/opentripplanner/transit/raptor/rangeraptor/standard/heuristics/HeuristicSearch.java
@@ -4,8 +4,8 @@ import java.util.Collection;
 import org.opentripplanner.transit.raptor.api.path.Path;
 import org.opentripplanner.transit.raptor.api.response.StopArrivals;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTripSchedule;
-import org.opentripplanner.transit.raptor.api.view.Heuristics;
-import org.opentripplanner.transit.raptor.api.view.Worker;
+import org.opentripplanner.transit.raptor.rangeraptor.internalapi.Heuristics;
+import org.opentripplanner.transit.raptor.rangeraptor.internalapi.Worker;
 
 /**
  * Combine Heuristics and Worker into one class to be able to retrieve the heuristics after the
